requirements.txt
# Computer Vision and Image Processing
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0

# Face Recognition and Emotion Detection
deepface>=0.0.79
tf-keras>=2.15.0
tensorflow>=2.15.0

# GUI Framework
# tkinter comes with Python standard library

# System Control
pyautogui>=0.9.54
pyttsx3>=2.90

# Windows COM API (PowerPoint Integration - Windows only)
comtypes>=1.2.0; sys_platform == 'win32'

# Logging (built-in with Python)
